basicClient/requirements.txt
flask
flask-cors
kafka-python
flask_sockets
websockets